Safari-611.1.10/LayoutTests/fast/forms/ValidityState-valueMissing-002-expected.txt

This test checks validity.valueMissing with some values or options with some values selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS valueMissingFor("input") is false
PASS valueMissingFor("textarea") is false
PASS valueMissingFor("select-with-placeholder") is false
PASS valueMissingFor("select-without-placeholder") is false
PASS valueMissingFor("select-with-fake-placeholder-size2") is false
PASS valueMissingFor("select-without-fake-placeholder-size2") is false
PASS valueMissingFor("select-with-fake-placeholder-multiple") is false
PASS valueMissingFor("select-without-fake-placeholder-multiple") is false
PASS valueMissingFor("select-with-fake-placeholder-size2-multiple") is false
PASS valueMissingFor("select-without-fake-placeholder-size2-multiple") is false
Updating valueMissing state by a key input:
PASS valueMissingFor("select-selecting-by-key") is true
PASS select.value is "a"
PASS valueMissingFor("select-selecting-by-key") is false
PASS valueMissingFor("select-selecting-by-key-2") is true
PASS select.value is "a"
PASS valueMissingFor("select-selecting-by-key-2") is false
PASS successfullyParsed is true

TEST COMPLETE
